Added from_contact_method field to Comment.
Now, when a comment is generated by incoming email,
comment.from_contact_method will be set to that email address. When
people receive a notification for that comment, it will look like it
came from that agent, so they will have the option to reply privately.

In the following cases, we do not have a "from email address" for
notifications:
1. Notification about something other than a comment
2. Comment was written on the website, so there is no associated
from_contact_method
3. Agent has permissions set so that the notification recipient cannot
view the agent's email contact method.

In these cases, the notification email will appear to come from the
relevant item itself. So if an action notice is about "Document 1", then
the email will come from "Deme notifier" <123@deme.stanford.edu>, and
replying to this email will generate a public comment on this item.

The upshot is, when the notification email appears to come from the
creator of the comment, then when you press "Reply" it will send them
back a private email, but if you press "Reply All" it will send a
private email to the creator AND a public comment to Deme. When the
notification email appears to come from an item, the only option is
"Reply", which will clearly send a public comment to Deme.